+++ backend/packages/eks/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cat-factory/eks
 
+## 0.1.248
+
+### Patch Changes
+
+- Updated dependencies [16576d6]
+  - @cat-factory/kernel@0.253.0
+  - @cat-factory/contracts@0.254.0
+  - @cat-factory/integrations@0.137.1
+
 ## 0.1.247
 
 ### Patch Changes